Execute chunks in parallel
Gone back to <set> instead of <unordered_set>.
For some reason, <unordered_set> causes some autosolves to need 6 guesses
(even when executing sequentially i.e. it is not a multithreading bug).
I have changed the unit test re a fixed code to use a code that needed
6 guesses when using <unordered_set>. It would be nice to understand
what the problem is with <unordered_set>.